backend/requirements.txt
requests
firebase-admin
beautifulsoup4
PyYAML
